+++ src/App.tsx
@@ -414,7 +414,7 @@ export const App: React.FC = () => {
           >
             {store.output.map((out, idx) => {
               return (
-                <Text fontSize="sm" mt={idx === 0 ? 2 : 0}>
+                <Text fontSize="sm" mt={idx === 0 ? 2 : 0} key={`output-${idx}`}>
                   <Text variant="gradient" as="span" fontWeight="bold">
                     [{idx + 1}]
                   </Text>{" "}
