+++ src/index.tsx
@@ -7,7 +7,7 @@ import * as serviceWorker from "./serviceWorker";
 import customTheme from "./theme";
 
 import StoreProvider from "./Store";
-import Tour from "./Tour";
+import Tour from "./tour";
 
 const container = document.getElementById("root");
 if (!container) throw new Error("Failed to find the root element");
